+++ docs/guides/live-video/client-setup/_category_.json
@@ -1,5 +1,5 @@
 {
   "position": 5,
-  "label": "Client Setup",
+  "label": "Live Video Client Setup",
   "collapsible": true
 }
