+++ docs/guides/livestream/client-setup/_category_.json
@@ -1,5 +1,5 @@
 {
   "position": 4,
-  "label": "Client Setup",
+  "label": "Livestream Client Setup",
   "collapsible": true
 }
