+++ docs/guides/voice-conf/client-setup/_category_.json
@@ -1,5 +1,5 @@
 {
   "position": 4,
-  "label": "Client Setup",
+  "label": "Voice Conference Client Setup",
   "collapsible": true
 }
