+++ package.json
@@ -32,7 +32,6 @@
     "@docusaurus/preset-classic": "^3.7.0",
     "@docusaurus/remark-plugin-npm2yarn": "^3.7.0",
     "@docusaurus/theme-live-codeblock": "^3.7.0",
-    "@docusaurus/theme-mermaid": "3.7.0",
     "@dytesdk/docs-ai-react": "^2.0.1",
     "@dytesdk/react-ui-kit": "^1.66.0",
     "@dytesdk/react-web-core": "file:modules/dytesdk-react-web-core-1.36.4-stripped.1.tgz",
@@ -54,6 +53,7 @@
     "zustand": "^4.5.2"
   },
   "devDependencies": {
+    "@docusaurus/theme-mermaid": "^3.9.1",
     "@docusaurus/tsconfig": "3.7.0",
     "@styled-icons/boxicons-logos": "^10.47.0",
     "@swc/core": "^1.10.11",
@@ -77,8 +77,8 @@
     "stream-browserify": "^3.0.0",
     "swc-loader": "^0.2.6",
     "tailwindcss": "^3.4.17",
-    "typedoc": "^0.23.28",
-    "typedoc-plugin-markdown": "^3.14.0",
+    "typedoc": "^0.24.8",
+    "typedoc-plugin-markdown": "^3.17.1",
     "typescript": "^5.0.2",
     "url": "^0.11.4"
   },
